requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
